Moved validate types to hours app. This app will also have to be redone soon
